Fixed bugs

+++ databaseAPI/tables/submissionsTable.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Float
 from databaseAPI.database import Base
 from datetime import datetime
 
@@ -9,7 +9,7 @@ class Submissions(Base):
     Id = Column(Integer, primary_key=True)
     UserId = Column(Integer, ForeignKey('users.Id'), nullable=ForeignKey)
     AddressId = Column(Integer, ForeignKey('walletaddress.Id'), nullable=ForeignKey)
-    Amount = Column(Integer, nullable=False)
+    Amount = Column(Float, nullable=False)
     TypeTrans = Column(String(255), nullable=False)  # crypto-rub | rub-crypto | crypto-crypto
     AddressUser = Column(String(255), nullable=False)
     Status = Column(String(255), default='WAIT')  # WAIT | ACCEPTED | COMPLETED
